[ODS-5914] Refactor XsdMetadata endpoints for route changes  (#808)
* Remove XsdMetadataFileMiddleware from solution

* Add xsd file endpoint to XsdMetadataController

* Remove XsdMetadataFileMiddleware tests

* Prevent possible ArgumentNullException
